requirements.txt
gunicorn==20.0.4
gunicorn==20.0.4
flask==1.1.2
google-api-core==1.17.0
google-api-python-client==1.8.4
google-auth==1.16.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1